Moves check_scoring_method to StructureScore.py from ScoreCache.py
