+++ pgmpy/estimators/GES.py
@@ -19,9 +19,8 @@ from pgmpy.estimators import (
     LogLikelihoodGauss,
     StructureEstimator,
     StructureScore,
+    check_scoring_method,
 )
-from pgmpy.estimators.ScoreCache import check_scoring_method
-
 from pgmpy.global_vars import logger
