+++ pgmpy/estimators/ScoreCache.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-from pgmpy.estimators import StructureScore
 from pgmpy.estimators import (
     AIC,
     BIC,
@@ -12,6 +11,7 @@ from pgmpy.estimators import (
     BICGauss,
     LogLikelihoodCondGauss,
     LogLikelihoodGauss,
+    StructureScore,
 )
 
 
@@ -122,51 +122,3 @@ class LRUCache:
             link[_PREV] = last
             link[_NEXT] = tail
         return value
-
-
-def check_scoring_method(scoring_method, data, use_cache):
-    supported_methods = {
-        "k2": K2,
-        "bdeu": BDeu,
-        "bds": BDs,
-        "bic-d": BIC,
-        "aic-d": AIC,
-        "ll-g": LogLikelihoodGauss,
-        "aic-g": AICGauss,
-        "bic-g": BICGauss,
-        "ll-cg": LogLikelihoodCondGauss,
-        "aic-cg": AICCondGauss,
-        "bic-cg": BICCondGauss,
-    }
-
-    if isinstance(scoring_method, str):
-        if scoring_method.lower() in [
-            "k2score",
-            "bdeuscore",
-            "bdsscore",
-            "bicscore",
-            "aicscore",
-        ]:
-            raise ValueError(
-                f"The scoring method names have been changed. Please refer the documentation."
-            )
-        elif scoring_method.lower() not in list(supported_methods.keys()):
-            raise ValueError(
-                f"Unknown scoring method. Please refer documentation for a list of supported score metrics."
-            )
-    elif not isinstance(scoring_method, StructureScore):
-        raise ValueError(
-            "scoring_method should either be one of k2score, bdeuscore, bicscore, bdsscore, aicscore, or an instance of StructureScore"
-        )
-
-    if isinstance(scoring_method, str):
-        score = supported_methods[scoring_method.lower()](data=data)
-    else:
-        score = scoring_method
-
-    if use_cache:
-        score_c = ScoreCache(score, data)
-    else:
-        score_c = score
-
-    return score, score_c
